+++ tests/test_friction.c
@@ -13,11 +13,11 @@ int test_sliding_friction(void)
         int ret = 0;
         int freq = 1000; // Using a high frequency to minimize integration error over the steps
         float v0 = 5.0f;
-        
+
         default_world(&w, freq);
         w.surface_count = 1;
         w.surfaces = &m;
-        
+
         // A large flat surface at y = 0
         m.vertex_count = 3;
         m.vertices = malloc(m.vertex_count * sizeof(struct vertex));
@@ -25,7 +25,7 @@ int test_sliding_friction(void)
         m.vertices[1].pos = (struct vec3) { .a = { -10.0f, 0.0f, 100.0f } };
         m.vertices[2].pos = (struct vec3) { .a = { 100.0f, 0.0f, -10.0f } };
         m.index_count = 3;
-        m.indices = malloc(m.index_count * sizeof(int));
+        m.indices = malloc(m.index_count * sizeof(uint16_t));
         m.indices[0] = 0;
         m.indices[1] = 1;
         m.indices[2] = 2;
@@ -48,22 +48,22 @@ int test_sliding_friction(void)
         o.dynamic_mu = 0.5f;
         o.steady_state = 0;
         o.contact_mesh = NULL;
-        
+
         // Expected distance: d = v^2 / (2 * mu * g)
         // Note: mu = sqrt(obj_mu * mesh_mu) = 0.5
         float expected_d = (v0 * v0) / (2.0f * 0.5f * KM_PHYS_G);
-        
+
         // Expected time: t = v / (mu * g)
         float expected_t = v0 / (0.5f * KM_PHYS_G);
-        
+
         int max_steps = 10000;
         int step = 0;
-        
+
         for (step = 0; step < max_steps; step++)
         {
                 update_object(step, &w, &o);
                 if (step % 50 == 0) {
-                        printf("step: %d, x: %f, y: %f, vx: %f, vy: %f, vabs: %f\n", 
+                        printf("step: %d, x: %f, y: %f, vx: %f, vy: %f, vabs: %f\n",
                                 step, o.p.p.x, o.p.p.y, o.p.v.x, o.p.v.y,
                                 vec3_dot(o.p.v, o.p.v));
                 }
@@ -76,24 +76,24 @@ int test_sliding_friction(void)
                         break;
                 }
         }
-        
+
         float actual_d = o.p.p.x;
         float actual_t = (float)step * w.dt;
-        
+
         // With numerical integration and discrete steps, allow small variations
         float epsilon_d = 0.02f;
         float epsilon_t = 0.05f;
-        
+
         if (fabsf(actual_d - expected_d) > epsilon_d)
         {
-                printf("Distance mismatch: got %f, expected %f (err %f)\n", 
+                printf("Distance mismatch: got %f, expected %f (err %f)\n",
                        actual_d, expected_d, fabsf(actual_d - expected_d));
                 ret = 1;
         }
-        
+
         if (fabsf(actual_t - expected_t) > epsilon_t)
         {
-                printf("Time mismatch: got %f, expected %f (err %f)\n", 
+                printf("Time mismatch: got %f, expected %f (err %f)\n",
                        actual_t, expected_t, fabsf(actual_t - expected_t));
                 ret = 1;
         }
@@ -101,7 +101,7 @@ int test_sliding_friction(void)
         // Clean up
         free(m.vertices);
         free(m.indices);
-        
+
         return ret;
 }
 
@@ -109,7 +109,7 @@ int main(void)
 {
         int fail = 0;
         fail = fail || test_sliding_friction();
-        
+
         if (fail) {
                 printf("test_friction failed\n");
         } else {
